reinforce_bot/requirements.txt
tensorflow-cpu==2.10
tensorflow-directml-plugin